@@ -1,11 +1,12 @@
 {
   "name": "whatsapp-web.js",
-  "version": "1.14.1",
+  "version": "1.18.0-alpha.1",
   "description": "Library for interacting with the WhatsApp Web API ",
   "main": "./index.js",
   "typings": "./index.d.ts",
   "scripts": {
-    "test": "mocha tests",
+    "test": "mocha tests --recursive --timeout 5000",
+    "test-single": "mocha",
     "shell": "node --experimental-repl-await ./shell.js",
     "generate-docs": "node_modules/.bin/jsdoc --configure .jsdoc.json --verbose"
   },
@@ -26,28 +27,34 @@
   "bugs": {
     "url": "https://github.com/pedroslopez/whatsapp-web.js/issues"
   },
-  "homepage": "https://guide.wwebjs.dev/",
+  "homepage": "https://wwebjs.dev/",
   "dependencies": {
     "@pedroslopez/moduleraid": "^5.0.2",
     "fluent-ffmpeg": "^2.1.2",
     "jsqr": "^1.3.1",
-    "mime": "^2.4.5",
+    "mime": "^3.0.0",
+    "node-fetch": "^2.6.5",
     "node-webpmux": "^3.1.0",
-    "puppeteer": "^10.1.0",
-    "sharp": "^0.28.3"
+    "puppeteer": "^13.0.0"
   },
   "devDependencies": {
-    "@types/node-fetch": "^2.5.11",
+    "@types/node-fetch": "^2.5.12",
     "chai": "^4.3.4",
-    "dotenv": "^10.0.0",
-    "eslint": "^7.27.0",
-    "eslint-plugin-mocha": "^9.0.0",
+    "chai-as-promised": "^7.1.1",
+    "dotenv": "^16.0.0",
+    "eslint": "^8.4.1",
+    "eslint-plugin-mocha": "^10.0.3",
     "jsdoc": "^3.6.4",
     "jsdoc-baseline": "^0.1.5",
     "mocha": "^9.0.2",
-    "sinon": "^11.1.1"
+    "sinon": "^13.0.1"
   },
   "engines": {
     "node": ">=12.0.0"
+  },
+  "optionalDependencies": {
+    "archiver": "^5.3.1",
+    "fs-extra": "^10.1.0",
+    "unzipper": "^0.10.11"
   }
 }
