@@ -2,24 +2,17 @@
  * ==== wwebjs-shell ====
  * Used for quickly testing library features
  * 
- * Running `npm run shell` will start WhatsApp Web in headless mode 
+ * Running `npm run shell` will start WhatsApp Web with headless=false
  * and then drop you into Node REPL with `client` in its context. 
  */
 
 const repl = require('repl');
-const fs = require('fs');
 
-const { Client } = require('./index');
-
-const SESSION_FILE_PATH = './session.json';
-let sessionCfg;
-if (fs.existsSync(SESSION_FILE_PATH)) {
-    sessionCfg = require(SESSION_FILE_PATH);
-}
+const { Client, LocalAuth } = require('./index');
 
 const client = new Client({
     puppeteer: { headless: false }, 
-    session: sessionCfg 
+    authStrategy: new LocalAuth()
 });
 
 console.log('Initializing...');
@@ -28,6 +21,10 @@ client.initialize();
 
 client.on('qr', () => {
     console.log('Please scan the QR code on the browser.');
+});
+
+client.on('authenticated', (session) => {
+    console.log(JSON.stringify(session));
 });
 
 client.on('ready', () => {
