@@ -108,7 +108,7 @@ class Contact extends Base {
          * @type {boolean}
          */
         this.isBlocked = data.isBlocked;
-
+        
         return super._patch(data);
     }
 
@@ -120,6 +120,22 @@ class Contact extends Base {
         return await this.client.getProfilePicUrl(this.id._serialized);
     }
 
+    /**
+     * Returns the contact's formatted phone number, (12345678901@c.us) => (+1 (234) 5678-901)
+     * @returns {Promise<string>}
+     */
+    async getFormattedNumber() {
+        return await this.client.getFormattedNumber(this.id._serialized);
+    }
+    
+    /**
+     * Returns the contact's countrycode, (1541859685@c.us) => (1)
+     * @returns {Promise<string>}
+     */
+    async getCountryCode() {
+        return await this.client.getCountryCode(this.id._serialized);
+    }
+    
     /**
      * Returns the Chat that corresponds to this Contact. 
      * Will return null when getting chat for currently logged in user.
@@ -167,13 +183,22 @@ class Contact extends Base {
      */
     async getAbout() {
         const about = await this.client.pupPage.evaluate(async (contactId) => {
-            return window.Store.Wap.statusFind(contactId);
+            const wid = window.Store.WidFactory.createWid(contactId);
+            return window.Store.StatusUtils.getStatus(wid);
         }, this.id._serialized);
 
         if (typeof about.status !== 'string')
             return null;
 
         return about.status;
+    }
+
+    /**
+     * Gets the Contact's common groups with you. Returns empty array if you don't have any common group.
+     * @returns {Promise<WAWebJS.ChatId[]>}
+     */
+    async getCommonGroups() {
+        return await this.client.getCommonGroups(this.id._serialized);
     }
     
 }
