@@ -19,13 +19,14 @@ class Message extends Base {
     }
 
     _patch(data) {
+        this._data = data;
+        
         /**
          * MediaKey that represents the sticker 'ID'
          * @type {string}
          */
         this.mediaKey = data.mediaKey;
-
-
+        
         /**
          * ID that represents the message
          * @type {object}
@@ -50,7 +51,7 @@ class Message extends Base {
          */
         this.body = this.hasMedia ? data.caption || '' : data.body || '';
 
-        /** 
+        /**
          * Message type
          * @type {MessageTypes}
          */
@@ -70,9 +71,9 @@ class Message extends Base {
 
         /**
          * ID for who this message is for.
-         * 
+         *
          * If the message is sent by the current user, it will be the Chat to which the message is being sent.
-         * If the message is sent by another user, it will be the ID for the current user. 
+         * If the message is sent by another user, it will be the ID for the current user.
          * @type {string}
          */
         this.to = (typeof (data.to) === 'object' && data.to !== null) ? data.to._serialized : data.to;
@@ -87,8 +88,8 @@ class Message extends Base {
          * String that represents from which device type the message was sent
          * @type {string}
          */
-        this.deviceType = data.id.id.length > 21 ? 'android' : data.id.id.substring(0,2) =='3A' ? 'ios' : 'web';
-        
+        this.deviceType = data.id.id.length > 21 ? 'android' : data.id.id.substring(0, 2) == '3A' ? 'ios' : 'web';
+
         /**
          * Indicates if the message was forwarded
          * @type {boolean}
@@ -114,14 +115,14 @@ class Message extends Base {
          * @type {boolean}
          */
         this.isStarred = data.star;
-        
+
         /**
          * Indicates if the message was a broadcast
          * @type {boolean}
          */
         this.broadcast = data.broadcast;
 
-        /** 
+        /**
          * Indicates if the message was sent by the current user
          * @type {boolean}
          */
@@ -132,6 +133,12 @@ class Message extends Base {
          * @type {boolean}
          */
         this.hasQuotedMsg = data.quotedMsg ? true : false;
+
+        /**
+         * Indicates the duration of the message in seconds
+         * @type {string}
+         */
+        this.duration = data.duration ? data.duration : undefined;
 
         /**
          * Location information contained in the message, if the message is type "location"
@@ -157,7 +164,7 @@ class Message extends Base {
             fromId: data.from._serialized,
             toId: data.to._serialized
         } : undefined;
-         
+
         /**
          * Indicates the mentions in the message body.
          * @type {Array<string>}
@@ -178,6 +185,18 @@ class Message extends Base {
          * @type {string}
          */
         this.token = data.token ? data.token : undefined;
+
+        /** 
+         * Indicates whether the message is a Gif
+         * @type {boolean}
+         */
+        this.isGif = Boolean(data.isGif);
+
+        /**
+         * Indicates if the message will disappear after it expires
+         * @type {boolean}
+         */
+        this.isEphemeral = data.isEphemeral;
 
         /** Title */
         if (data.title) {
@@ -202,7 +221,7 @@ class Message extends Base {
         /**
          * Links included in the message.
          * @type {Array<{link: string, isSuspicious: boolean}>}
-         * 
+         *
          */
         this.links = data.links;
 
@@ -210,7 +229,7 @@ class Message extends Base {
         if (data.dynamicReplyButtons) {
             this.dynamicReplyButtons = data.dynamicReplyButtons;
         }
-        
+
         /** Selected Button Id **/
         if (data.selectedButtonId) {
             this.selectedButtonId = data.selectedButtonId;
@@ -220,7 +239,7 @@ class Message extends Base {
         if (data.listResponse && data.listResponse.singleSelectReply.selectedRowId) {
             this.selectedRowId = data.listResponse.singleSelectReply.selectedRowId;
         }
-        
+
         return super._patch(data);
     }
 
@@ -228,6 +247,32 @@ class Message extends Base {
         return this.fromMe ? this.to : this.from;
     }
 
+    /**
+     * Reloads this Message object's data in-place with the latest values from WhatsApp Web. 
+     * Note that the Message must still be in the web app cache for this to work, otherwise will return null.
+     * @returns {Promise<Message>}
+     */
+    async reload() {
+        const newData = await this.client.pupPage.evaluate((msgId) => {
+            const msg = window.Store.Msg.get(msgId);
+            if(!msg) return null;
+            return window.WWebJS.getMessageModel(msg);
+        }, this.id._serialized);
+
+        if(!newData) return null;
+        
+        this._patch(newData);
+        return this;
+    }
+
+    /**
+     * Returns message in a raw format
+     * @type {Object}
+     */
+    get rawData() {
+        return this._data;
+    }
+    
     /**
      * Returns the Chat this message was sent in
      * @returns {Promise<Chat>}
@@ -268,12 +313,12 @@ class Message extends Base {
     }
 
     /**
-     * Sends a message as a reply to this message. If chatId is specified, it will be sent 
-     * through the specified Chat. If not, it will send the message 
+     * Sends a message as a reply to this message. If chatId is specified, it will be sent
+     * through the specified Chat. If not, it will send the message
      * in the same Chat as the original message was sent.
-     * 
-     * @param {string|MessageMedia|Location} content 
-     * @param {string} [chatId] 
+     *
+     * @param {string|MessageMedia|Location} content
+     * @param {string} [chatId]
      * @param {MessageSendOptions} [options]
      * @returns {Promise<Message>}
      */
@@ -291,16 +336,30 @@ class Message extends Base {
     }
 
     /**
+     * React to this message with an emoji
+     * @param {string} reaction - Emoji to react with. Send an empty string to remove the reaction.
+     * @return {Promise}
+     */
+    async react(reaction){
+        await this.client.pupPage.evaluate(async (messageId, reaction) => {
+            if (!messageId) { return undefined; }
+            
+            const msg = await window.Store.Msg.get(messageId);
+            await window.Store.sendReactionToMsg(msg, reaction);
+        }, this.id._serialized, reaction);
+    }
+
+    /**
      * Accept Group V4 Invite
      * @returns {Promise<Object>}
      */
     async acceptGroupV4Invite() {
         return await this.client.acceptGroupV4Invite(this.inviteV4);
     }
-    
+
     /**
-     * Forwards this message to another chat
-     * 
+     * Forwards this message to another chat (that you chatted before, otherwise it will fail)
+     *
      * @param {string|Chat} chat Chat model or chat ID to which the message will be forwarded
      * @returns {Promise}
      */
@@ -330,7 +389,7 @@ class Message extends Base {
             if (msg.mediaData.mediaStage != 'RESOLVED') {
                 // try to resolve media
                 await msg.downloadMedia({
-                    downloadEvenIfExpensive: true, 
+                    downloadEvenIfExpensive: true,
                     rmrReason: 1
                 });
             }
@@ -350,13 +409,14 @@ class Message extends Base {
                     type: msg.type,
                     signal: (new AbortController).signal
                 });
-    
-                const data = window.WWebJS.arrayBufferToBase64(decryptedMedia);
-    
+
+                const data = await window.WWebJS.arrayBufferToBase64Async(decryptedMedia);
+
                 return {
                     data,
                     mimetype: msg.mimetype,
-                    filename: msg.filename
+                    filename: msg.filename,
+                    filesize: msg.size
                 };
             } catch (e) {
                 if(e.status && e.status === 404) return undefined;
@@ -365,7 +425,7 @@ class Message extends Base {
         }, this.id._serialized);
 
         if (!result) return undefined;
-        return new MessageMedia(result.mimetype, result.data, result.filename);
+        return new MessageMedia(result.mimetype, result.data, result.filename, result.filesize);
     }
 
     /**
@@ -376,8 +436,8 @@ class Message extends Base {
         await this.client.pupPage.evaluate((msgId, everyone) => {
             let msg = window.Store.Msg.get(msgId);
 
-            if (everyone && msg.id.fromMe && msg.canRevoke()) {
-                return window.Store.Cmd.sendRevokeMsgs(msg.chat, [msg], true);
+            if (everyone && msg.id.fromMe && msg._canRevoke()) {
+                return window.Store.Cmd.sendRevokeMsgs(msg.chat, [msg], {type: 'Sender'});
             }
 
             return window.Store.Cmd.sendDeleteMsgs(msg.chat, [msg], true);
@@ -392,7 +452,7 @@ class Message extends Base {
             let msg = window.Store.Msg.get(msgId);
 
             if (msg.canStar()) {
-                return msg.chat.sendStarMsgs([msg], true);
+                return window.Store.Cmd.sendStarMsgs(msg.chat, [msg], false);
             }
         }, this.id._serialized);
     }
@@ -405,7 +465,7 @@ class Message extends Base {
             let msg = window.Store.Msg.get(msgId);
 
             if (msg.canStar()) {
-                return msg.chat.sendStarMsgs([msg], false);
+                return window.Store.Cmd.sendUnstarMsgs(msg.chat, [msg], false);
             }
         }, this.id._serialized);
     }
@@ -428,14 +488,10 @@ class Message extends Base {
     async getInfo() {
         const info = await this.client.pupPage.evaluate(async (msgId) => {
             const msg = window.Store.Msg.get(msgId);
-            if(!msg) return null;
-            
-            return await window.Store.Wap.queryMsgInfo(msg.id);
-        }, this.id._serialized);
+            if (!msg) return null;
 
-        if(info.status) {
-            return null;
-        }
+            return await window.Store.MessageInfo.sendQueryMsgInfo(msg);
+        }, this.id._serialized);
 
         return info;
     }
@@ -446,9 +502,9 @@ class Message extends Base {
      */
     async getOrder() {
         if (this.type === MessageTypes.ORDER) {
-            const result = await this.client.pupPage.evaluate((orderId, token) => {
-                return window.WWebJS.getOrderDetail(orderId, token);
-            }, this.orderId, this.token);
+            const result = await this.client.pupPage.evaluate((orderId, token, chatId) => {
+                return window.WWebJS.getOrderDetail(orderId, token, chatId);
+            }, this.orderId, this.token, this._getChatId());
             if (!result) return undefined;
             return new Order(this.client, result);
         }
