@@ -17,5 +17,6 @@ module.exports = {
     Call: require('./Call'),
     Buttons: require('./Buttons'),
     List: require('./List'),
-    Payment: require('./Payment')
+    Payment: require('./Payment'),
+    Reaction: require('./Reaction'),
 };
