@@ -5,40 +5,64 @@ exports.ExposeStore = (moduleRaidStr) => {
     eval('var moduleRaid = ' + moduleRaidStr);
     // eslint-disable-next-line no-undef
     window.mR = moduleRaid();
-    window.Store = window.mR.findModule('Chat')[0].default;
-    window.Store.AppState = window.mR.findModule('STREAM')[0].default;
-    window.Store.Conn = window.mR.findModule('Conn')[0].default;
+    window.Store = Object.assign({}, window.mR.findModule(m => m.default && m.default.Chat)[0].default);
+    window.Store.AppState = window.mR.findModule('Socket')[0].Socket;
+    window.Store.Conn = window.mR.findModule('Conn')[0].Conn;
+    window.Store.BlockContact = window.mR.findModule('blockContact')[0];
+    window.Store.Call = window.mR.findModule('CallCollection')[0].CallCollection;
+    window.Store.Cmd = window.mR.findModule('Cmd')[0].Cmd;
     window.Store.CryptoLib = window.mR.findModule('decryptE2EMedia')[0];
-    window.Store.Wap = window.mR.findModule('Wap')[0].default;
-    window.Store.SendSeen = window.mR.findModule('sendSeen')[0];
-    window.Store.SendClear = window.mR.findModule('sendClear')[0];
-    window.Store.SendDelete = window.mR.findModule('sendDelete')[0];
-    window.Store.genId = window.mR.findModule('randomId')[0].default;
-    window.Store.SendMessage = window.mR.findModule('addAndSendMsgToChat')[0];
-    window.Store.MsgKey = window.mR.findModule((module) => module.default && module.default.fromString)[0].default;
+    window.Store.DownloadManager = window.mR.findModule('downloadManager')[0].downloadManager;
+    window.Store.MDBackend = window.mR.findModule('isMDBackend')[0].isMDBackend();
+    window.Store.Features = window.mR.findModule('FEATURE_CHANGE_EVENT')[0].LegacyPhoneFeatures;
+    window.Store.GroupMetadata = window.mR.findModule((module) => module.default && module.default.handlePendingInvite)[0].default;
     window.Store.Invite = window.mR.findModule('sendJoinGroupViaInvite')[0];
-    window.Store.OpaqueData = window.mR.findModule(module => module.default && module.default.createFromData)[0].default;
+    window.Store.InviteInfo = window.mR.findModule('sendQueryGroupInvite')[0];
+    window.Store.Label = window.mR.findModule('LabelCollection')[0].LabelCollection;
     window.Store.MediaPrep = window.mR.findModule('MediaPrep')[0];
     window.Store.MediaObject = window.mR.findModule('getOrCreateMediaObject')[0];
-    window.Store.MediaUpload = window.mR.findModule('uploadMedia')[0];
-    window.Store.Cmd = window.mR.findModule('Cmd')[0].default;
+    window.Store.NumberInfo = window.mR.findModule('formattedPhoneNumber')[0];
     window.Store.MediaTypes = window.mR.findModule('msgToMediaType')[0];
-    window.Store.VCard = window.mR.findModule('vcardFromContactModel')[0];
+    window.Store.MediaUpload = window.mR.findModule('uploadMedia')[0];
+    window.Store.MsgKey = window.mR.findModule((module) => module.default && module.default.fromString)[0].default;
+    window.Store.MessageInfo = window.mR.findModule('sendQueryMsgInfo')[0];
+    window.Store.OpaqueData = window.mR.findModule(module => module.default && module.default.createFromData)[0].default;
+    window.Store.QueryExist = window.mR.findModule('queryExists')[0].queryExists;
+    window.Store.QueryProduct = window.mR.findModule('queryProduct')[0];
+    window.Store.QueryOrder = window.mR.findModule('queryOrder')[0];
+    window.Store.SendClear = window.mR.findModule('sendClear')[0];
+    window.Store.SendDelete = window.mR.findModule('sendDelete')[0];
+    window.Store.SendMessage = window.mR.findModule('addAndSendMsgToChat')[0];
+    window.Store.SendSeen = window.mR.findModule('sendSeen')[0];
+    window.Store.User = window.mR.findModule('getMaybeMeUser')[0];
+    window.Store.UploadUtils = window.mR.findModule((module) => (module.default && module.default.encryptAndUpload) ? module.default : null)[0].default;
     window.Store.UserConstructor = window.mR.findModule((module) => (module.default && module.default.prototype && module.default.prototype.isServer && module.default.prototype.isUser) ? module.default : null)[0].default;
     window.Store.Validators = window.mR.findModule('findLinks')[0];
+    window.Store.VCard = window.mR.findModule('vcardFromContactModel')[0];
+    window.Store.Wap = window.mR.findModule('queryLinkPreview')[0].default;
     window.Store.WidFactory = window.mR.findModule('createWid')[0];
-    window.Store.BlockContact = window.mR.findModule('blockContact')[0];
-    window.Store.GroupMetadata = window.mR.findModule((module) => module.default && module.default.handlePendingInvite)[0].default;
-    window.Store.Sticker = window.mR.findModule('Sticker')[0].default.Sticker;
-    window.Store.UploadUtils = window.mR.findModule((module) => (module.default && module.default.encryptAndUpload) ? module.default : null)[0].default;
-    window.Store.Label = window.mR.findModule('LabelCollection')[0].default;
-    window.Store.Features = window.mR.findModule('FEATURE_CHANGE_EVENT')[0].default;
-    window.Store.QueryOrder = window.mR.findModule('queryOrder')[0];
-    window.Store.QueryProduct = window.mR.findModule('queryProduct')[0];
-    window.Store.DownloadManager = window.mR.findModule('DownloadManager')[0].default;
-    window.Store.Call = window.mR.findModule('CallCollection')[0].default;
+    window.Store.ProfilePic = window.mR.findModule('profilePicResync')[0];
+    window.Store.PresenceUtils = window.mR.findModule('sendPresenceAvailable')[0];
+    window.Store.ChatState = window.mR.findModule('sendChatStateComposing')[0];
+    window.Store.GroupParticipants = window.mR.findModule('sendPromoteParticipants')[0];
+    window.Store.JoinInviteV4 = window.mR.findModule('sendJoinGroupViaInviteV4')[0];
+    window.Store.findCommonGroups = window.mR.findModule('findCommonGroups')[0].findCommonGroups;
+    window.Store.StatusUtils = window.mR.findModule('setMyStatus')[0];
+    window.Store.ConversationMsgs = window.mR.findModule('loadEarlierMsgs')[0];
+    window.Store.sendReactionToMsg = window.mR.findModule('sendReactionToMsg')[0].sendReactionToMsg;
+    window.Store.createOrUpdateReactionsModule = window.mR.findModule('createOrUpdateReactions')[0];
+    window.Store.StickerTools = {
+        ...window.mR.findModule('toWebpSticker')[0],
+        ...window.mR.findModule('addWebpMetadata')[0]
+    };
+  
+    window.Store.GroupUtils = {
+        ...window.mR.findModule('sendCreateGroup')[0],
+        ...window.mR.findModule('sendSetGroupSubject')[0],
+        ...window.mR.findModule('markExited')[0]
+    };
 
-    if(!window.Store.Chat._find) {
+    if (!window.Store.Chat._find) {
         window.Store.Chat._find = e => {
             const target = window.Store.Chat.get(e);
             return target ? Promise.resolve(target) : Promise.resolve({
@@ -51,14 +75,6 @@ exports.ExposeStore = (moduleRaidStr) => {
 exports.LoadUtils = () => {
     window.WWebJS = {};
 
-    window.WWebJS.getNumberId = async (id) => {
-
-        let result = await window.Store.Wap.queryExist(id);
-        if (result.jid === undefined)
-            throw 'The number provided is not a registered whatsapp user';
-        return result.jid;
-    };
-
     window.WWebJS.sendSeen = async (chatId) => {
         let chat = window.Store.Chat.get(chatId);
         if (chat !== undefined) {
@@ -68,14 +84,14 @@ exports.LoadUtils = () => {
         return false;
 
     };
-    
+
     window.WWebJS.sendMessage = async (chat, content, options = {}) => {
         let attOptions = {};
         if (options.attachment) {
-            attOptions = options.sendMediaAsSticker 
+            attOptions = options.sendMediaAsSticker
                 ? await window.WWebJS.processStickerData(options.attachment)
                 : await window.WWebJS.processMediaData(options.attachment, {
-                    forceVoice: options.sendAudioAsVoice, 
+                    forceVoice: options.sendAudioAsVoice,
                     forceDocument: options.sendMediaAsDocument,
                     forceGif: options.sendVideoAsGif
                 });
@@ -85,7 +101,6 @@ exports.LoadUtils = () => {
             delete options.attachment;
             delete options.sendMediaAsSticker;
         }
-
         let quotedMsgOptions = {};
         if (options.quotedMessageId) {
             let quotedMessage = window.Store.Msg.get(options.quotedMessageId);
@@ -145,25 +160,29 @@ exports.LoadUtils = () => {
 
         if (options.linkPreview) {
             delete options.linkPreview;
-            const link = window.Store.Validators.findLink(content);
-            if (link) {
-                const preview = await window.Store.Wap.queryLinkPreview(link.url);
-                preview.preview = true;
-                preview.subtype = 'url';
-                options = { ...options, ...preview };
+
+            // Not supported yet by WhatsApp Web on MD
+            if(!window.Store.MDBackend) {
+                const link = window.Store.Validators.findLink(content);
+                if (link) {
+                    const preview = await window.Store.Wap.queryLinkPreview(link.url);
+                    preview.preview = true;
+                    preview.subtype = 'url';
+                    options = { ...options, ...preview };
+                }
             }
         }
         
-        let extraOptions = {};
+        let buttonOptions = {};
         if(options.buttons){
             let caption;
-            if(options.buttons.type === 'chat') {
+            if (options.buttons.type === 'chat') {
                 content = options.buttons.body;
                 caption = content;
-            }else{
+            } else {
                 caption = options.caption ? options.caption : ' '; //Caption can't be empty
             }
-            extraOptions = {
+            buttonOptions = {
                 productHeaderImageRejected: false,
                 isFromTemplate: false,
                 isDynamicReplyButtonsMsg: true,
@@ -176,12 +195,12 @@ exports.LoadUtils = () => {
             delete options.buttons;
         }
 
+        let listOptions = {};
         if(options.list){
             if(window.Store.Conn.platform === 'smba' || window.Store.Conn.platform === 'smbi'){
                 throw '[LT01] Whatsapp business can\'t send this yet';
             }
-            extraOptions = {
-                ...extraOptions,
+            listOptions = {
                 type: 'list',
                 footer: options.list.footer,
                 list: {
@@ -191,31 +210,48 @@ exports.LoadUtils = () => {
                 body: options.list.description
             };
             delete options.list;
-            delete extraOptions.list.footer;
+            delete listOptions.list.footer;
         }
-                
+
+        const meUser = window.Store.User.getMaybeMeUser();
+        const isMD = window.Store.MDBackend;
+
         const newMsgId = new window.Store.MsgKey({
-            fromMe: true,
-            remote: chat.id,
-            id: window.Store.genId(),
+            from: meUser,
+            to: chat.id,
+            id: window.Store.MsgKey.newId(),
+            participant: isMD && chat.id.isGroup() ? meUser : undefined,
+            selfDir: 'out',
         });
+
+        const extraOptions = options.extraOptions || {};
+        delete options.extraOptions;
+
+        const ephemeralSettings = {
+            ephemeralDuration: chat.isEphemeralSettingOn() ? chat.getEphemeralSetting() : undefined,
+            ephemeralSettingTimestamp: chat.getEphemeralSettingTimestamp() || undefined,
+            disappearingModeInitiator: chat.getDisappearingModeInitiator() || undefined,
+        };
 
         const message = {
             ...options,
             id: newMsgId,
             ack: 0,
             body: content,
-            from: window.Store.Conn.wid,
+            from: meUser,
             to: chat.id,
             local: true,
             self: 'out',
             t: parseInt(new Date().getTime() / 1000),
             isNewMsg: true,
             type: 'chat',
+            ...ephemeralSettings,
             ...locationOptions,
             ...attOptions,
             ...quotedMsgOptions,
             ...vcardOptions,
+            ...buttonOptions,
+            ...listOptions,
             ...extraOptions
         };
 
@@ -223,11 +259,25 @@ exports.LoadUtils = () => {
         return window.Store.Msg.get(newMsgId._serialized);
     };
 
+    window.WWebJS.toStickerData = async (mediaInfo) => {
+        if (mediaInfo.mimetype == 'image/webp') return mediaInfo;
+
+        const file = window.WWebJS.mediaInfoToFile(mediaInfo);
+        const webpSticker = await window.Store.StickerTools.toWebpSticker(file);
+        const webpBuffer = await webpSticker.arrayBuffer();
+        const data = window.WWebJS.arrayBufferToBase64(webpBuffer);
+
+        return {
+            mimetype: 'image/webp',
+            data
+        };
+    };
+
     window.WWebJS.processStickerData = async (mediaInfo) => {
         if (mediaInfo.mimetype !== 'image/webp') throw new Error('Invalid media type');
 
         const file = window.WWebJS.mediaInfoToFile(mediaInfo);
-        let filehash = await window.WWebJS.getFileHash(file);	
+        let filehash = await window.WWebJS.getFileHash(file);
         let mediaKey = await window.WWebJS.generateHash(32);
 
         const controller = new AbortController();
@@ -313,10 +363,11 @@ exports.LoadUtils = () => {
 
     window.WWebJS.getMessageModel = message => {
         const msg = message.serialize();
-        
+
+        msg.isEphemeral = message.isEphemeral;
         msg.isStatusV3 = message.isStatusV3;
-        msg.links = (message.getLinks()).map(link => ({ 
-            link: link.href, 
+        msg.links = (message.getLinks()).map(link => ({
+            link: link.href,
             isSuspicious: Boolean(link.suspiciousCharacters && link.suspiciousCharacters.size)
         }));
 
@@ -326,24 +377,30 @@ exports.LoadUtils = () => {
         if (msg.dynamicReplyButtons) {
             msg.dynamicReplyButtons = JSON.parse(JSON.stringify(msg.dynamicReplyButtons));
         }
-        if(msg.replyButtons) {
+        if (msg.replyButtons) {
             msg.replyButtons = JSON.parse(JSON.stringify(msg.replyButtons));
         }
-        
+
+        if (typeof msg.id.remote === 'object') {
+            msg.id = Object.assign({}, msg.id, { remote: msg.id.remote._serialized });
+        }
+
         delete msg.pendingAckUpdate;
-        
+
         return msg;
     };
 
 
     window.WWebJS.getChatModel = async chat => {
+
         let res = chat.serialize();
         res.isGroup = chat.isGroup;
         res.formattedTitle = chat.formattedTitle;
         res.isMuted = chat.mute && chat.mute.isMuted;
 
         if (chat.groupMetadata) {
-            await window.Store.GroupMetadata.update(chat.id._serialized);
+            const chatWid = window.Store.WidFactory.createWid((chat.id._serialized));
+            await window.Store.GroupMetadata.update(chatWid);
             res.groupMetadata = chat.groupMetadata.serialize();
         }
 
@@ -361,7 +418,7 @@ exports.LoadUtils = () => {
     };
 
     window.WWebJS.getChats = async () => {
-        const chats = window.Store.Chat.models;
+        const chats = window.Store.Chat.getModelsArray();
 
         const chatPromises = chats.map(chat => window.WWebJS.getChatModel(chat));
         return await Promise.all(chatPromises);
@@ -393,12 +450,12 @@ exports.LoadUtils = () => {
     };
 
     window.WWebJS.getContacts = () => {
-        const contacts = window.Store.Contact.models;
+        const contacts = window.Store.Contact.getModelsArray();
         return contacts.map(contact => window.WWebJS.getContactModel(contact));
     };
 
     window.WWebJS.mediaInfoToFile = ({ data, mimetype, filename }) => {
-        const binaryData = atob(data);
+        const binaryData = window.atob(data);
 
         const buffer = new ArrayBuffer(binaryData.length);
         const view = new Uint8Array(buffer);
@@ -415,15 +472,29 @@ exports.LoadUtils = () => {
 
     window.WWebJS.arrayBufferToBase64 = (arrayBuffer) => {
         let binary = '';
-        const bytes = new Uint8Array( arrayBuffer );
+        const bytes = new Uint8Array(arrayBuffer);
         const len = bytes.byteLength;
         for (let i = 0; i < len; i++) {
-            binary += String.fromCharCode( bytes[ i ] );
+            binary += String.fromCharCode(bytes[i]);
         }
-        return window.btoa( binary );
+        return window.btoa(binary);
     };
 
-    window.WWebJS.getFileHash = async (data) => {                  
+    window.WWebJS.arrayBufferToBase64Async = (arrayBuffer) =>
+        new Promise((resolve, reject) => {
+            const blob = new Blob([arrayBuffer], {
+                type: 'application/octet-stream',
+            });
+            const fileReader = new FileReader();
+            fileReader.onload = () => {
+                const [, data] = fileReader.result.split(',');
+                resolve(data);
+            };
+            fileReader.onerror = (e) => reject(e);
+            fileReader.readAsDataURL(blob);
+        });
+
+    window.WWebJS.getFileHash = async (data) => {
         let buffer = await data.arrayBuffer();
         const hashBuffer = await crypto.subtle.digest('SHA-256', buffer);
         return btoa(String.fromCharCode(...new Uint8Array(hashBuffer)));
@@ -433,7 +504,7 @@ exports.LoadUtils = () => {
         var result = '';
         var characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
         var charactersLength = characters.length;
-        for ( var i = 0; i < length; i++ ) {
+        for (var i = 0; i < length; i++) {
             result += characters.charAt(Math.floor(Math.random() * charactersLength));
         }
         return result;
@@ -458,15 +529,18 @@ exports.LoadUtils = () => {
     };
 
     window.WWebJS.sendChatstate = async (state, chatId) => {
+        if (window.Store.MDBackend) {
+            chatId = window.Store.WidFactory.createWid(chatId);
+        }
         switch (state) {
         case 'typing':
-            await window.Store.Wap.sendChatstateComposing(chatId);
+            await window.Store.ChatState.sendChatStateComposing(chatId);
             break;
         case 'recording':
-            await window.Store.Wap.sendChatstateRecording(chatId);
+            await window.Store.ChatState.sendChatStateRecording(chatId);
             break;
         case 'stop':
-            await window.Store.Wap.sendChatstatePaused(chatId);
+            await window.Store.ChatState.sendChatStatePaused(chatId);
             break;
         default:
             throw 'Invalid chatstate';
@@ -478,12 +552,12 @@ exports.LoadUtils = () => {
     window.WWebJS.getLabelModel = label => {
         let res = label.serialize();
         res.hexColor = label.hexColor;
-        
+
         return res;
     };
 
     window.WWebJS.getLabels = () => {
-        const labels = window.Store.Label.models;
+        const labels = window.Store.Label.getModelsArray();
         return labels.map(label => window.WWebJS.getLabelModel(label));
     };
 
@@ -497,8 +571,9 @@ exports.LoadUtils = () => {
         return (chat.labels || []).map(id => window.WWebJS.getLabel(id));
     };
 
-    window.WWebJS.getOrderDetail = async (orderId, token) => {
-        return window.Store.QueryOrder.queryOrder(orderId, 80, 80, token);
+    window.WWebJS.getOrderDetail = async (orderId, token, chatId) => {
+        const chatWid = window.Store.WidFactory.createWid(chatId);
+        return window.Store.QueryOrder.queryOrder(chatWid, orderId, 80, 80, token);
     };
 
     window.WWebJS.getProductMetadata = async (productId) => {
@@ -510,21 +585,4 @@ exports.LoadUtils = () => {
 
         return undefined;
     };
-};
-
-exports.MarkAllRead = () => {
-    let Chats = window.Store.Chat.models;
-
-    for (let chatIndex in Chats) {
-        if (isNaN(chatIndex)) {
-            continue;
-        }
-
-        let chat = Chats[chatIndex];
-
-        if (chat.unreadCount > 0) {
-            chat.markSeen();
-            window.Store.Wap.sendConversationSeen(chat.id, chat.getLastMsgKeyForAction(), chat.unreadCount - chat.pendingSeenCount);
-        }
-    }
 };
