@@ -1,20 +1,17 @@
 'use strict';
 
-const sharp = require('sharp');
 const path = require('path');
 const Crypto = require('crypto');
 const { tmpdir } = require('os');
 const ffmpeg = require('fluent-ffmpeg');
 const webp = require('node-webpmux');
 const fs = require('fs').promises;
-
 const has = (o, k) => Object.prototype.hasOwnProperty.call(o, k);
 
 /**
  * Utility methods
  */
 class Util {
-
     constructor() {
         throw new Error(`The ${this.constructor.name} class may not be instantiated.`);
     }
@@ -23,7 +20,7 @@ class Util {
         var result = '';
         var characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
         var charactersLength = characters.length;
-        for ( var i = 0; i < length; i++ ) {
+        for (var i = 0; i < length; i++) {
             result += characters.charAt(Math.floor(Math.random() * charactersLength));
         }
         return result;
@@ -55,33 +52,19 @@ class Util {
      * 
      * @returns {Promise<MessageMedia>} media in webp format
      */
-    static async formatImageToWebpSticker(media) {
+    static async formatImageToWebpSticker(media, pupPage) {
         if (!media.mimetype.includes('image'))
             throw new Error('media is not a image');
-      
+
         if (media.mimetype.includes('webp')) {
             return media;
         }
-      
-        const buff = Buffer.from(media.data, 'base64');
-      
-        let sharpImg = sharp(buff);
-        sharpImg = sharpImg.webp();
-      
-        sharpImg = sharpImg.resize(512, 512, {
-            fit: 'contain',
-            background: { r: 0, g: 0, b: 0, alpha: 0 },
-        });
-      
-        let webpBase64 = (await sharpImg.toBuffer()).toString('base64');
-      
-        return {             
-            mimetype: 'image/webp',
-            data: webpBase64,
-            filename: media.filename,
-        };
+
+        return pupPage.evaluate((media) => {
+            return window.WWebJS.toStickerData(media);
+        }, media);
     }
-      
+
     /**
      * Formats a video to webp
      * @param {MessageMedia} media
@@ -91,14 +74,14 @@ class Util {
     static async formatVideoToWebpSticker(media) {
         if (!media.mimetype.includes('video'))
             throw new Error('media is not a video');
-      
+
         const videoType = media.mimetype.split('/')[1];
 
         const tempFile = path.join(
             tmpdir(),
             `${Crypto.randomBytes(6).readUIntLE(0, 6).toString(36)}.webp`
         );
-      
+
         const stream = new (require('stream').Readable)();
         const buffer = Buffer.from(
             media.data.replace(`data:${media.mimetype};base64,`, ''),
@@ -135,17 +118,17 @@ class Util {
                 .toFormat('webp')
                 .save(tempFile);
         });
-      
+
         const data = await fs.readFile(tempFile, 'base64');
         await fs.unlink(tempFile);
-      
-        return {             
+
+        return {
             mimetype: 'image/webp',
             data: data,
             filename: media.filename,
         };
     }
-      
+
     /**
      * Sticker metadata.
      * @typedef {Object} StickerMetadata
@@ -161,14 +144,14 @@ class Util {
      * 
      * @returns {Promise<MessageMedia>} media in webp format
      */
-    static async formatToWebpSticker(media, metadata) {
+    static async formatToWebpSticker(media, metadata, pupPage) {
         let webpMedia;
 
-        if (media.mimetype.includes('image')) 
-            webpMedia = await this.formatImageToWebpSticker(media);
-        else if (media.mimetype.includes('video')) 
+        if (media.mimetype.includes('image'))
+            webpMedia = await this.formatImageToWebpSticker(media, pupPage);
+        else if (media.mimetype.includes('video'))
             webpMedia = await this.formatVideoToWebpSticker(media);
-        else 
+        else
             throw new Error('Invalid media format');
 
         if (metadata.name || metadata.author) {
