@@ -1,13 +1,24 @@
 const path = require('path');
-const Client = require('../src/Client');
-const Util = require('../src/util/Util');
+const { Client, LegacySessionAuth, LocalAuth } = require('..');
 
 require('dotenv').config();
 
 const remoteId = process.env.WWEBJS_TEST_REMOTE_ID;
 if(!remoteId) throw new Error('The WWEBJS_TEST_REMOTE_ID environment variable has not been set.');
 
+function isUsingLegacySession() {
+    return Boolean(process.env.WWEBJS_TEST_SESSION || process.env.WWEBJS_TEST_SESSION_PATH);
+}
+
+function isMD() {
+    return Boolean(process.env.WWEBJS_TEST_MD);
+}
+
+if(isUsingLegacySession() && isMD()) throw 'Cannot use legacy sessions with WWEBJS_TEST_MD=true';
+
 function getSessionFromEnv() {
+    if (!isUsingLegacySession()) return null;
+
     const envSession = process.env.WWEBJS_TEST_SESSION;
     if(envSession) return JSON.parse(envSession);
 
@@ -16,17 +27,28 @@ function getSessionFromEnv() {
         const absPath = path.resolve(process.cwd(), envSessionPath);
         return require(absPath);
     }
-    
-    throw new Error('No session found in environment.');
 }
 
-function createClient({withSession, options: additionalOpts}={}) {
+function createClient({authenticated, options: additionalOpts}={}) {
     const options = {};
-    if(withSession) {
-        options.session = getSessionFromEnv();
+
+    if(authenticated) {
+        const legacySession = getSessionFromEnv();
+        if(legacySession) {
+            options.authStrategy = new LegacySessionAuth({
+                session: legacySession
+            });
+        } else {
+            const clientId = process.env.WWEBJS_TEST_CLIENT_ID;
+            if(!clientId) throw new Error('No session found in environment.');
+            options.authStrategy = new LocalAuth({
+                clientId
+            });
+        }
     }
 
-    return new Client(Util.mergeDefault(options, additionalOpts || {}));
+    const allOpts = {...options, ...(additionalOpts || {})};
+    return new Client(allOpts);
 }
 
 function sleep(ms) {
@@ -36,5 +58,7 @@ function sleep(ms) {
 module.exports = {
     sleep, 
     createClient,
-    remoteId
+    isUsingLegacySession,
+    isMD,
+    remoteId,
 };
